[DAD-566] feat: 비디오에 컴파운드 패턴 적용
